hello integers

+++ Escape/src/model/backbone/building/elements/Exit.java
@@ -17,7 +17,7 @@ public class Exit {
 		end = b;
 	}
 	
-	public Exit(double x1, double y1, double x2, double y2){
+	public Exit(int x1, int y1, int x2, int y2){
 		begin = new Point(x1, y1);
 		end = new Point(x2, y2);
 	}
